2_graspdiffusion_baseline/requirements.txt
numpy
tqdm
h5py
scikit_learn==1.3.0
gorilla-core==0.2.7.8
# torch==2.0.1
scipy==1.10.1
trimesh==4.0.7
opencv-contrib-python
git+https://github.com/openai/CLIP.git

timm
open3d
# pytorch3d